fix: Similar _id field in aggregate queries.
