+++ src/extend-aggregate.js
@@ -1,6 +1,7 @@
 'use strict';
 
 const generateKey = require('./generate-key');
+const recoverObjectId = require('./recover-objectid');
 let hasBeenExtended = false;
 
 module.exports = function(mongoose, cache) {
@@ -28,7 +29,9 @@ module.exports = function(mongoose, cache) {
 
       return new Promise((resolve, reject) => {
         cache.get(key, (err, cachedResults) => { //eslint-disable-line handle-callback-err
-          if (cachedResults) {
+          if (cachedResults != null) {
+            cachedResults = recoverObjectId(mongoose, cachedResults);
+
             callback(null, cachedResults);
             return resolve(cachedResults);
           }
